requirements.txt
dill>=0.3.9
matplotlib>=3.9.2
numpy>=1.26.4
pandas>=2.2.3
pillow>=11.0.0
plotly>=5.24.1
scikit-learn>=1.5.2
seaborn>=0.13.2
shap>=0.46.0
streamlit>=1.40.1
joblib>=1.3.2
altair>=5.3.0
